Create the UpsertTransaction Azure Function and implement the insert logic

+++ src/AKExpnesesTracker/AKExpensesTracker.Server.Data/Models/Transaction.cs
@@ -30,7 +30,7 @@ namespace AKExpensesTracker.Server.Data.Models
 		public bool IsIncome { get; private set; }
 
 		[JsonProperty("amount")]
-		public decimal Amount { get; private set; }
+		public double Amount { get; private set; }
 
 		[JsonProperty("creationDate")]
 		public DateTime CreationDate { get; private set; }
@@ -67,7 +67,7 @@ namespace AKExpensesTracker.Server.Data.Models
 
 		public static Transaction Create(string walletId, 
 										 string userId, 
-										 decimal amount,
+										 double amount,
 										 string category,
 										 bool isIncome,
 										 string? description = null,
@@ -89,7 +89,7 @@ namespace AKExpensesTracker.Server.Data.Models
 		}
 
 		public void Update(bool isIncome,
-						   decimal amount,
+						   double amount,
 						   string category,
 						   string? description = null,
 						   string[]? tags = null,
